requirements.txt
prometheus-client==0.12.0
pydigitemp==1.4.1
